requirements.txt
genanki==0.13.1
customtkinter==5.2.2
pyyaml==6.0.3
